python/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
astor==0.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
distro==1.9.0
duckdb==1.0.0
et-xmlfile==1.1.0
Faker==19.13.0
fastapi==0.112.2
fonttools==4.53.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
Jinja2==3.1.4
jiter==0.5.0
kiwisolver==1.4.5
loguru==0.7.2
MarkupSafe==2.1.5
matplotlib==3.9.2
numpy==1.26.4
openai==1.43.0
openpyxl==3.1.5
packaging==24.1
pandas==1.5.3
pandasai==2.2.14
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
sqlglot==25.18.0
sqlglotrs==0.2.9
starlette==0.38.4
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
win32-setctime==1.1.0
